libGLES: remove need for egl.cfg support.
starting in K, the need for egl.cfg and so the naming of the
vendor libGLES library is no longer necessary.  instead, the
system attempts to locate the vendor libGLES from its location
on the filesystem and the name corresponding to the default
name.

note: we keep the named library '_nexus' because of the direct
dependency in gralloc, the aosp framework can still fallback to
this name.

refs #SWANDROID-895

Change-Id: I0518e728674ea376ea1df213877f44b7a396104f
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>