SWANDROID-897: [L] Cannot mount USB stick
Updates to init.rc and fstab.bcm_platform to force vold to mount
the sdcard using the "vdc" command, as the MountService in
frameworks has been set to only support emulated external storage
(i.e. bind mounting the internal "/data/media" dir to "/storage/sdcard").

When we move to the full Android-L sources, then we won't need this
workaround, and can instead create a "storage_list.xml" file with the
appropriate settings to indicate that we have physical external
removable storage.

refs #SWANDROID-897

Change-Id: I7a53ac3eb337420042bea2a7109a2d2178b438a0