SWANDROID-905: [L] Prevent Android from entering standby due to no user interaction after 1 minute
Rather than having to force the user to enable "Developer Options" and to click
on the "stay awake" checkbox to prevent the STB from entering standby after 1 minute,
this fix automatically enables the "stay awake" feature.  It does so by using the
"svc" command to inform the PowerManagerService to never enter sleep if the user is not
interacting with the STB.

The good news is that if the user presses the POWER button, the STB still enters standby.

refs #SWANDROID-905

Change-Id: I74d5d6fb2cfec700b258584b856732645ad22e0c