generated: add default module.
though generated, some modules are directly
invoked in some android.mk which cause a build
failure on android makefile walk through if no
version of the file is present.  it was expected
that those could be removed, but too earger of a
cleanup.  possibly a solution would be to make
those .inc a dependency in the makefile including
them rather than a pure include, that way the build
would only fail at the time of build which would
have created the generated module.

refs #SWANDROID-857

Change-Id: Ie1f124eab6fc7b39c3ea864c7ccd6d069d576e2a
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>