[L] Audio HAL using new interfaces
Current Broadcom audio HAL uses the Android legacy audio interfaces. However,
support for the legacy interfaces in Android L PDK is limited and unstable.
It is necesssary to switch to the new Android audio interfaces.

The new implementation also attempts to separate out the operations that are
implementaion specific, to allow future support for USB audio devices.

refs #SWANDROID-915

Change-Id: I498a768a1000923436bbde21be17677917d775f8