SWANDROID-932: [L] Support HDMI CEC functionality through the HAL
New hdmi_cec HAL module introduced that uses the NexusCecService
to pass CEC and HDMI commands to/from Nexus.

I've updated the NexusCecService to properly handle sending CEC
commands with an "initiator" source address and for waiting for
the message to be sent before returning.

Many changes, but HDMI CEC now works well.  For example, if you
press the AV button on a Sony TV and select your current HDMI
source, it will show the source with OSD text at the top saying
it is a 7XXXX TEST STB.

We can also send CEC commands manually using the send_cec utility.

I've also introduced the listener/observer design pattern to allow
a separate process (or same) to listen for certain events.  In
this case, the HDMI CEC control service listens for HDMI hotplug
and HDMI CEC message events from NexusService and reacts to them
accordingly.

Currently, Google have set the HDMI CEC device type for STBs to be
0x04 which is of type "playback".  Normally, STBs with a tuner
should be set to 0x03 which is of type "tuner".  We shall await
Google's comments on this oversight.

refs #SWANDROID-932

Change-Id: I7b9b5568b768742bb68e59a26e00378676c1ed69