device: adding script to generate target.
this is the script to be used to generate the
portion of the device configuration which is dependent
on the hardware targetted and which requires proper
refsw setup.

the script is based off the refsw plat one and so any
hardware known to the given refsw version being used
can be configured/selected.

refs #SWANDROID-857

Change-Id: Iccc7272d7ea85b8c7aae38c50b31bdaf85e3d82b
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>