Android-L: Add support of UnpackAndroid tools and scripts
The proper way to load the compiled Android image to target platform is via fastboot, and doing so will properly preserve the access permission of the file system.  However, if flashing the image via fastboot is not available for whatever reasons, the UnpackAndroid tools can be used instead to stage the image and the corresponding filesystems with the proper access permission preserved.

refs #SWANDROID-987

Change-Id: I5598eaff5d8ccb24e6cbcae956da675c0763f453