SWANDROID-1001: [L] Wrong HDMI colorspace chosen at boot time causing a pinkish tint to be observed.
When the "-ignore_edid" option is passed to nxserver, the pink tint is observed.  This is a
refsw issue, but we need not be using this option any longer if the graphics rendering is
set to 1080p and the default/preferred HDMI video resolution is also 1080p.  In this case,
the graphics won't be upscaled in the GFD and will look crisp/sharp.

In the past, the graphics resolution used to be 720p and if we did not ignore the EDID
information, then the default HDMI output resolution would typically be 1080p causing
the GFD to upscale and the end result was blurry looking graphics.

refs #SWANDROID-1001

Change-Id: Ife3671e4ac5c05583e96dff90dbd1810b1e172bc