SWANDROID-1001: [L] Wrong HDMI colorspace chosen at boot time causing a pinkish tint to be observed
Backed out compile time changes to determine whether the HDMI EDID information should be ignored
or not depending on display resolution.   Instead, the determination is done at run-time by
examining a property.

So, if the desired output resolution from the STB is set to 1920x1080, then the EDID information
from the TV is *NOT* ignored, as the preferred TV resolution is also typically 1920x1080p.
In this case, the graphics feeder (GFD) will not perform any scaling operations to cause the
graphics quality to be reduced due to scaling artefacts/filtering.

If the desired output resolution from the STB is not 1080p, then the EDID information from the
connected TV is ignored and the output resolution of the TV is then made to match the same
resoution of the graphics/video output from the STB.  This ensures that no scaling is performed
in the GFD to detract from the output graphics quality.

Because we are not ingoring the EDID information in the 1080p case, and that for some platforms
the default graphics output resolution is also 1080p, then we get no pink screen tint and the
best possible graphics quality due to no scaling.

refs #SWANDROID-1001

Change-Id: Ie9c5e3a6a706d26ff1123dde57f61384ef691107