Fixed dummy audio input driver to be CTS compatible
Apparently the dummy audio input driver has its buffer size initialized
to be zero which confuses AudioFlinger when it tries to read samples.
Now a valid buffer size is initialized in the dummy driver so now
CTS would not complain about not reading any input samples.

Also changed the sampling frequency from 8K to 44.1K as 44.1K is tested
in CTS.

refs #SWANDROID-989

Change-Id: I3b16c613a22d979ec9e56f0ba87680713c3825e2