device: disguise nexus node under for cts compliance.
use a node that is already part of the exception list to
disguise the coalesced nexus node.  add the ueventd entry
which will be used once the SWANDROID-1023 is implemented.

note that the node chosen works out okay and if we ever end
up with a mali based gpu design, nexus is likely to need to
change as well anyways.

refs #SWANDROID-1026

Change-Id: Id88e76e1ff5a56ec99e9ad830d1daa889b72c9b4
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>