device: disguise nexus-wake node for cts compliance.
use a node that is already part of the exception list to
disguise the nexus-wake node.  add the ueventd entry
which will be used once the SWANDROID-1023 is implemented.

note that the node chosen works out okay and if we ever end
up with a mali based gpu design, nexus is likely to need to
change as well anyways.

refs #SWANDROID-1026

Change-Id: I429a360e33aba42b89c80b048f7a254feea80148