omx-nexus-decoder: add property to control 4k selection.
refs #SWANDROID-1099

do not force asking for 4K decode when selecting h265 decoder.
doing so and given the 'randomized' selection of decoder from
omx-android down to nexus, as documented on #SWANDROID-1098,
may lead to picking the wrong decoder which would not offer sufficient
memory to allow proper decoding and which would itself lead to crash
possibly (as documented on #SWANDROID-1096).

to avoid those problems, we add a property to determine whether
to force 4K decoder when selecting h265, the property defaults to
'0' (or false), meaning we will not ask for 4K, but can be changed
run time for testing otherwise via setprop. ie: 'adb shell setprop
bcm.omx.nx.dec.h265.ask4k 1'.

this can be removed (or defaulted back to true) once #SWANDROID-1098
and #SWANDROID-1096 are both addressed properly.

Change-Id: Ic917a6cc2179b29999ebb532dbbd2692850b4f1c
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>