SWANDROID-963: [L] Support Power Management modes
We now insmod the wakeup driver with the appropriate device name
before we insmod the nexus driver.  Note, wakeup works fine on the
l-release-latest Android branch because the reference software is
latest and has the necessary fixes for the following JIRA's:

1) http://jira.broadcom.com/browse/SWANDROID-1023
2) http://jira.broadcom.com/browse/SWANDROID-1026

Please be aware that the l-release branch as of today won't allow
the STB to exit standby due to reference software being old.

The second change in nexus_power.cpp is not to turn on/off the TV
through HDMI CEC when exiting and entering standby respectively.
The reason is because Android-L now has its own HDMI CEC logic that
controls the behaviour of the STB (and connected devices).

refs #SWANDROID-963

Change-Id: Ieb909d1d0937e011c0c8d210e2f7814f6852059e