plat-droid: use kernel toolchains from prebuilt.
kernel toolchains are now part of the manifest, so use them
from there and no need to rely on installed ones on the
build system.

note: keep legacy path alive for a bit to pass jenkins
builder which is unable to deal properly with cross
dependencies, this will be removed at a later time.

refs #SWANDROID-857

Change-Id: I54ee88803623ce60fa7d135f93a8d1c3bb56bdc0
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>