ashmem: disguise device node for CTS compliance.
refs #SWANDROID-854

allow to disguise the device node for nx_ashmem for
certification compliance, using node that is already in the
exception list.

this is the simplest way to achieve compliance without needing
to rework the code nor ask for device exception, yet follows
the model already adopted for nexus nodes (nexus kernel driver
and nexus wake driver).

Change-Id: I1e65813651fe24177fc4003be0e2ace3a11b603a
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>