SWANDROID-963: [L] Support Power Management modes
I've now added the ability for a client to register for a standby monitor
callback.  This callback allows a client to decide whether or not it is
ready to enter standby or not.  If the client is not ready, then Nexus
will NOT enter standby until all clients are ready.  By default, all
clients will automatically acknowledge that they can enter standby.

Both HWC and the Audio HAL now register for a standby monitor callback
to help provide information to Nexus in deciding whether or not it
can enter standby successfully.

I've also re-factored our power.cpp code to clean it up.

refs #SWANDROID-963

Change-Id: Ic26f220b0038e4de1e7f73c0a8b569c4210593ae