Modification To Starup Code =========================== There were two issues related to Startup of the decoder.
1. The decoder was started at the time when the component was initialized.
This should be done when the component enters the idle state. Only the
required initialization, for example, connection to the nexus service
should be establised when the component is initialized.

2. When the connection to the nexus server was made, we were not setting
the required decoder type correctly. We have added this code now.

3. Apart from this there was a minor issue with saving the codec config
data. We were running out of the config data buffer. So we see some error
messages on the screen. We now have a logic to overwrite and re-use the
codec config data.

refs #SWANDROID-1098

Change-Id: I2342a5852c43e6be66ac16f5352f5fc0558f11c7