camera: Default /dev/video0 as front camera
The current camera hal implementation assumes that /dev/video0 is always
the rear camera, and /dev/video1 is always the front by default.  For
the STB situation, this is not the case, so treat the default camera
that comes up as a front camera.  This fixes cts android.app
testCameraFeatures when using a usb webcam.

Also added a check to the camera.cfg so that future camera devices could
be configured at run-time through camera.cfg rather than relying on the
default.  Fixed a bug with the camera.cfg parser code as the logic was
reversed.

refs #SWANDROID-1068

Change-Id: I8b0c731eb5dc16c8ce093d8f44ee7a08aa60b2c0