Fixing playpump event handling
There were two issues resolved with playpump event handling.
One is that if you have just shutdown and had a playpump event
callback pending you may get a callback when you have a NULL
playpump handle.  The other problem was that the total
descriptor count was not updated each place a descriptor was
submitted potentially leading to problems with flush or
EOS handling.

refs #SWANDROID-1352

Change-Id: I1fdef1aedbd83ccf9672a920c65a3940ec3c8188