Fixing EOS handling if multiple output buffers are pending
If multiple output buffers were pending and one contained EOS
there was an issue that would generate an assert reporting
"Additional frames received after EOS".  The check for having
already reported EOS was in the wrong place and needed to be checked
inside the processing loop instead of prior to entering it.

refs #SWANDROID-1351

Change-Id: Id85365ab3da0488e836394fb27ef8b862c913d07