Flushing frame buffer queue on shutdown
When shutting down in metadata mode, the outstanding gralloc buffers
are not known by the component during OMX_FreeBuffer.  In Native
mode, they are.  So, when the component shuts down in metadata
mode it must flush the allocated buffer list internally as opposed
to expecting calls to FreeBuffer to remove linked nodes.

refs #SWANDROID-1329

Change-Id: Id9ac4bd56769d5ba533cdb111d0a7741350b1699