Checking run status during return frame call
If frames are returned via IPC from HWC while the decoder is
in the idle state because of a flush or other action, we need
to check the run state of the decoder before trying to return
frames.  If the decoder is stopped our internal queue can get
out of sync permanently.

refs #SWANDROID-1329

Change-Id: Ia1cfdad67cc2d93b1da76cd9c698985d7b496926