BT/DHD: 43569USB support
For 43569USB support, the firmware needs to be loaded only once on
boot to enable shared USB port support, which allows a single USB
interface to be used to connect to both WLAN and Bluetooth
functionality.  This feature is denoted by "-vusb-" in the DHD
firmware configuration.

In addition, unloading the bcmdhd.ko module would also disrupt
the USB connectivity to Bluetooth.  This patch (along with change
to BoardConfigs.mk) changes the bcmdhd.ko to be loaded only once.

refs #SWANDROID-1339

Change-Id: I6ca3b394a933e3eeb544cf7b5a0f309ab5df10dc
Signed-off-by: Edwin Chan <echan@broadcom.com>