lmkstats: calculate the lmk's value configured.
to help understand the lmk related problems with the
tied memory configuration, add this small exe which
allows to calculate the settings that android would
impose on the lmk.

the main function is lifted directly from the framework
code since it is what is used to adjust the effective
oom values posted to kernel.

tweaks on the application values can be made to see the
effect on the oom scores posted in relation with the
error reported during run time by the oom-killer.

refs #SWANDROID-1359

Change-Id: I19356b6722e2d747503b968977b06fe85cdc4ff2
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>