SWANDROID-1343: [L] Feature: add Broadcom "doze" mode to allow STB to enter an intermediate power mode
Added a new Broadcom "doze" mode to allow the STB to enter an intermediate power state before
potentially transitioning to the final off power state.  The customer can define what this "doze"
state will be by a property (e.g. s1 or s2).  A doze timeout can also be specified via a property
and if it is "0", then the STB will by-pass the "doze" state and will enter the "off" power state
instead.   If the "doze" timeout is very large, then this can be used to prevent the STB from
entering the "off" power state.

Added a new Broadcom interactive timeout to allow the STB to actually enter the desired "off" state
depending on the value of a property.  This interactive timeout can be used to fine tune when the
STB should actually power off and is customer specific.  It is recommended that this timeout is of
a value equal to or greater than 1s to ensure that all Android framework components have received
the "ACTION_SCREEN_OFF" broadcast Intent and have taken action accordingly.

We now acquire a partial wake-lock on entry to S1 and S2 modes to prevent the auto-suspend
feature of Android placing the CPU in to deep sleep.

Also, cleaned up the code to make it more readable.

refs #SWANDROID-1343

Change-Id: I4ee42e6782fb4a8b607b082438b6de03bc1afd72