Fixing handling of EOS+Payload in the same buffer
There was a bug in the output frame processing loop that would
return any buffer with EOS marked as an empty frame with
nFilledLen=0.  This is incorrect if the frame has both EOS
and a valid picture in the same buffer.  This change will
look for an EOS-only frame and return nFilledLen=0 otherwise
it will process the frame as any other frame.

refs #SWANDROID-1351

Change-Id: I2a3adea5a102a6857d330c450fb85721effbb8cd