Fixing two OMX race conditions
This change fixes two race conditions found in the AdaptivePlayback
CTS suite.

The first bug is related to output buffer format
selection.  Previously we were latching the output buffer mode
on the component thread when transitioning to idle or enabling
the output port.  This is a race condition because the component
thread may not get scheduled before UseBuffer is called and an
error would result due to invalid buffer size.  The fix is to
correctly latch the state during UseBuffer/AllocBuffer on the
client's context.

The second bug was with the DisplayFrame callback from HWC during
a flush command.  If the input port flush invalidated a frame still
held in hwc, the DisplayFrame handler would still attempt to
display it afterward.  The DisplayFrame handler needed to check
the buffer state and ignore anything not ready to be returned to
nexus.

refs #SWANDROID-1329

Change-Id: Id8ab54a8b855d25a60b79857bab31530077987d0