hwc/sideband: allow creation of sideband clients.
allow creation and usage of sideband clients which do
not own the surface (nor window surface) they render into,
this allows the user of the sideband to control it instead,
yet leaving the geometry control into the hwc hands.

this is the (new) default mode for sideband windows, we may
want to make that tunable further on if need be.

refs #SWANDROID-1420

Change-Id: I99f6bb89b084c288d85f285a19871381953e8faf
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>