SWANDROID-1347: [L] Feature: When STB enters standby it should also place the TV in to standby via CEC
Added code to allow the STB to send the CEC <Standby> opcode to the TV to place the TV in to standby
when the STB is also placed in to standby.  This option is enabled by setting the following property:

persist.sys.hdmi.tx_standby_cec=1

Code has also been added to send the CEC <Image View On> and <Active Source> opcodes to the TV when the
STB is brought out of standby.  This allows the TV to exit standby and select the HDMI input that the
STB is connected to when the STB exits standby.  This option is enabled by setting the following property:

persist.sys.hdmi.tx_view_on_cec=1

The HDMI CEC code has also been updated to accommodate the "set_option" HAL method for HDMI_OPTION_WAKEUP
and HDMI_OPTION_SYSTEM_CEC_CONTROL options.  The former option is passed down from the Android HDMI CEC
framework code to inform Nexus whether the STB should be woken up upon receipt of a relevant CEC command.
The later option is used by the HDMI CEC HAL code to trigger the sending of the <Standby> and
<Image View On> CEC commands and also to decide whether the Android HDMI CEC framework code should be
allowed to send and receive CEC commands.

The CEC test apps have been updated to use the correct IPC API to determine whether HDMI Cec is enabled
or not.

The hotplug "connected" event is dispatched to listeners if and only if the HDMI device is connected
AND powered.

refs #SWANDROID-1347

Change-Id: I21c9f0fd9ac6f53d150f762caa6c67f4485e61a7