Refactoring PES insertion logic
This change refactors the PES insertion logic to achieve a few
goals.  First, config data will no longer be sent as a separate
PES packeet from frame payload.  Now, config data will be accumulated
and sent as part of the PES packet for the next frame.

The second part of the change is to switchover from potentialy
unbounded PES packets to splitting into multiple bounded PES packets
if the frame payload is too large for a single packet.  Codecs
like VP8 that do not have start code emulation in the bitstream
can potentially fool transport into thinking there is a fake PES
start code 00 00 01 in the bitstream and cause hardware to drop
frame payload.

refs #SWANDROID-1368

Change-Id: Ibe9bbd2c6c1c3195747b612595c2128c7e2911d3