hwc/sideband: do not allocate surface clients.
allocating surface clients on the hwc and having them
used in a different process for geometry+rendering control
isn't really working properly today and it also breaks one
of the fundamental asumptions of nexus architecture, so
let's not do that for now.  let the client user of the
sideband channel allocate and fully own/manage its surface
client.

refs #SWANDROID-1420

Change-Id: I9b8551cc6fbe85cbdeb12328a4bdacd9ec68539a
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>