hwc: do not allocate surface client for video planes.
turns out that for standard video planes, just like for
sidebands ones, we cannot allocate the surface client nor
the video surface in hwc since the owner needs to be the
same one than the owner of the decoder in order for things
to work out properly.

add a binder api to allow set/get of the geometry information
associated with the video window, hook up an event notification
when such changes to the current owner of the video window.

refs #SWANDROID-1420

Change-Id: I16e8b1120a52246b6c6cf53c22333ab8ffd302b8
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>