SWANDROID-1460: [L] l-release does not boot but shows black screen
This is a stack corruption problem!

"NEXUS_PlatformConfiguration" is being used inside methods as an
atomic variable but it does consume a large amount of stack space
to hold it.  I've changed the functions to use dynamically
allocated memory (BKNI_Malloc) instead to prevent the stack from
being corrupted.

I've also removed superflous declarations in some of the methods
and fixed up the HDMI and HDCP hotplug routines which were
badly written.

refs #SWANDROID-1460

Change-Id: I4a73db6b73db8d798f3e3e764415fe19e29c8527