SWANDROID-1462: [L] Prevent screen from dimming due to user inactivity
Part 2 of 2 changes:
===================
Removed the "disable-sleep" service that was forcing the STB to stay
awake when the device was being charged.  We no longer need this as
the user inactivity timeout is now set to just over 24 days.

The second change was the remove the now redundant boot_complete
trigger that was used by "disable-sleep" to know when to issue
the "svc power" command.

refs #SWANDROID-1462

Change-Id: I145df06ed6a6de7e02c6299979830f44fb56ac8a