SWANDROID-1456: [L] Black screen and hang after attempting to exit S1 standby
The hwc prepare function was being called in S1 standby before Nexus had
been taken out of standby into S0.  The call to
"NxClient_GetSurfaceClientComposition()" during S1 standby was hanging
causing any further NxClient calls in the same process to also deadlock
because of the global API LOCK in "nxclient_ipc.c" still being held.  This
was causing SurfaceFlinger to hang and the Watchdog then kicked in.
The solution was not exercise the "set" or "prepare" methods for the
primary display when entered or entering in to standby.

Both S1 and S2 standby have been tested with this change on a BCM97252C
platform with WiFi disabled and not further deadlocks have been observed.

refs #SWANDROID-1456

Change-Id: Ia0777a46c1d32e8f775e9f566769f3da882fae15