audio: Implement custom audio policy manager
To support Android Remote Control app from a mobile device, a
custom audio policy manager is required to allow the use of
the remote submix device as input device.  The following
build flags are also required:

   USE_LEGACY_AUDIO_POLICY := 0
   USE_CUSTOM_AUDIO_POLICY := 1

This code implements the same basic remote submix input selection
as the Nexus Player.

refs #SWANDROID-1279

Change-Id: If2a42355ca1869a559e00bd0131c6fd521d32329
Signed-off-by: Edwin Chan <echan@broadcom.com>