SWANDROID-1343: [L] Feature: add Broadcom "doze" mode to allow STB to enter an intermediate power mode
Now added the ability to set the doze timeout to a negative value to prevent the STB from
entering the off state.

When the STB was placed in S2 and a CEC command from the TV was sent to wake up the STB,
the STB tried to enter S3 mode rather than exiting S2.

Fixed seg fault when both doze timeout and verbose logging are enabled.

Added wakelock verbose debugging.

refs #SWANDROID-1343

Change-Id: Ia25a48e0506fda374a1b241cd3116af34d287d85