Enabling YCbCr Planar output in OMX
This switches the output color format from the OMX video decoder
to a 420 Planar YCbCr format.  It also provides a placeholder to
destripe the video data in gralloc_lock by embedding the striped
surface handle (unlocked for multiprocess) in the gralloc buffer.
Currently, planar destriping is not working in nexus so for now
the code will do a destripe to 422 and then sw-convert to 420
planar afterward.

refs #SWANDROID-1050

Change-Id: Ic094003c10fbff5e0ecf7c5c527f6ce3a6f971a8