SWANDROID-1477: [L] Random crashes on entry standby.
Fixed a nasty bug that was causing the PC to jump off to "hyperspace"
in the "StandbyMonitorThread" thread in nexus_nx_client.cpp.  The
problem was that the "mStandbyMonitorCallback" was being corrupted
and this was causing code execution to go beyond the virtual
address space of the current process resulting in a seg fault.

The corruption was caused by the fact that in nexus_egl_client.cpp
the "nexus_egl_client::IEGL_nexus_join" method was destroying
the NexusIPCClient instance that was holding both the
"mStandbyMonitorCallback" and "mStandbyMonitorContext" fields.

The fix is to pass in both the callback and context directly to
the "StandbyMonitorThread" constructor and not reference a
NexusIPCClient instance.  The fix now prevents the random
crashes on entry to standby.

refs #SWANDROID-1477

Change-Id: Ib6797f1fbc39f678a94fb27c843f90c73a49a55c