SWANDROID-1509: [L] Segmentation fault in libnexusservice
This fixes another subtle but nasty bug where the NexusServerContext
in nexusnxservice.cpp was not being instantiated properly due to
a duplicate definition in nexusservice.cpp.

The fix was to ensure that NexusServerContext in nexusnxservice.cpp
is renamed to NexusNxServerContext and that it inherits from
NexusServerContext in nexusservice.cpp.

The customer has confirmed that the patch fixes the problem that
was raised in IMS case 870423.

refs #SWANDROID-1509

Change-Id: I4ef90cb30440f16615f11dbcbba3fa8615025476