SWANDROID-1505: [L] Prevent HDMI and Display Nexus API calls during standby to avoid deadlocks
Initial changes to prevent the HDMI and HDCP callbacks from calling
in to Nexus (NxClient) during the standby state.  This prevents the
API's from blocking which can cause other non HDMI/Display NxClient
API's in the same process to also block.  The net effect is usually
a deadlock in the system on entering standby.

refs #SWANDROID-1505

Change-Id: I02ccc63a2675bee235d3a08a4ebc5ef315c946fd