hwc: increase max layer count to 10+1
it has been seen in some situation that we may need up
to 6+1 layer (the extra +1 is the floating HWC_FRAMEBUFFER_TARGET
added by SurfaceFlinger for GLES composition).

thus increase our layer count to 10+1 for safety.  we probably
want to shuffle some of those eventually but for now we do not
really use that much at once on the NSC side, so we could have a
floating index, for further study, this will do.

refs #SWANDROID-1516

Change-Id: Iefd7b19cf023e6f170ef30424ff4f60cd63238b5
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>