gralloc/hwc: push yv12 to yuv422 conversion to application scope.
reworked the conversion from yv12 to yuv422 to push it into the
scope of the application (gralloc) rather than at composition
(hwc).

fixed up the usage of g2d within gralloc since it has to be per
process client in order to be used within the scope of lock/unlock.

should also help with SWANDROID-1527

refs #SWANDROID-1536

Change-Id: If0a489b3dc13ce819ba813a6cac14478068ba109
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>