Adding resolution change support
Output port definition and output crop rectangle will now update
on a change in decoded data format.  In metadata output, the decoder
will continue to output after the change.  In other output modes, the
decoder will output will stop until the component is reconfigured.

This also changes the decoder output aspect ratio mode back to eFull
so that stretching is supported as per android expectations.  Prior to
this, the output would be letterbox and running tests like AdaptivePlayback
the SW decoder output was noticeably different from the HW decoder.

refs #SWANDROID-1555

Change-Id: I10da00655248928b0c59a4268cda7108156b22fe