PM: Spoof Power/Wakeup event when AON GPIOs get toggled
refs #SWANDROID-1567

When "Wake on BLE" or "Wake on WLAN" wakes the system up via toggling
the AON GPIO, the screen remains black because Android is not aware that
the screen needs to be turned on. We need to spoof a Power/Wakeup event
so that Android will turn the screen on for us.

Change-Id: I0c36fc15fdea25a5d3b0a59193c58a009aa764dd