Resized partitions to follow BoardConfig.mk
(This change was lost when migrating from 8GB to 16GB script)

Using Bash arithmetic to calculate partition start/end block numbers.
Note: the previous hardcoded partitions sizes were off by +1 block.
The last number in the -n option is the ending block (start+size-1),
not the first block of the next partition (start+size).

Aligned partitions on 1MB boundary.

Removed redundant type set for partition 1 (-t 1:0700). It was on the
command line of the recovery partition 6, but applied to partition 1.
(Should have been -t 6:0700 to apply to partition 6, but was not
necessary anyway.)

refs #SWANDROID-1298

Change-Id: I329877c29d2ab84022913d6b92480e94dd2742b2