Resolving race conditions with on-the-fly format changes
There were two issues with adaptive playback format changes.  One,
if the clip regions were updates on the fly, there is a chance that
the clip rectangle comes back out of sync with the frame itself.  This
will reject attempts to set an invalid clip rectangle.  The other issue
was if buffers returned during OMX_FreeBuffer as part of a port format change
came back out of order the frame queue could become out of sync with nexus and
assert.

refs #SWANDROID-1555

Change-Id: Idb0c0a59134bbe3ce6b880a5389e9a50ba78ab89