Fixed YV12 conversion in gralloc
There is a bug in gralloc_destripe_yv12() where the YUV values are
incorrectly copied from the packed YUYV because the "x" in the for loop
should be bounded by width instead of height. Luckily most of the
graphics/video are displayed in landscape where width is greater than
height. Otherwise memory corruption may happen for video displayed in
portrait.

refs #SWANDROID-1635

Change-Id: Ia7fa2b23c768f419a4f70f94eb831e9264fbe330