hwc: forced refresh of display.
this is a bit ugly...  but... for the time being,
we aren't properly handling the framebuffer target
refresh which would leave the display to always miss
the last frame update.

this hack up allows to query get a refresh going every
now and then which will also happens to bring the last
frame on display as it should and eliminate visual
artifacts.

refs #SWANDROID-1503

Change-Id: I85f5754f06bf71a6e8806bd23efb40b0e74e2833
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>