unpack: Fix broken rsync delete
UnpackAndroid was recently changed to
use the rsync delete option, instead
of just wiping the whole partition first.
This option did not work because a
wildcard (*) was used in the rsync command,
so files intended to be deleted were left.
Remove the wildcard since it is not needed,
and fix the delete functionality.

refs #SWANDROID-1537

Change-Id: Icf65e73ffaaf09bba9ba04a0bf79ecff74d8d307