Fix binary structure layout error.
Theres a struct NexusClientContext mismatch between nexusservice.cpp
and nexusnxservice.cpp In libnexusservice nexusservice.cpp adds a field
hdmiInput to the definition struct NexusClientContext that is not
in nexusnxservice.cpp.

As a result a call to  getNexusClient in
NexusNxService::createClientContext which falls back to
NexusService::getNexusClient will use a misplaced pid field.

refs #SWANDROID-1667

Change-Id: Id2709b104f3b1f5a24db93b7f1b1bdafeb49b8f3
Signed-off-by: Tomek Bury <tomek@broadcom.com>