SWANDROID-1645: [L] Cannot wakeup after powerdown from android refsw changes for Phase 2
The refsw now has the ability to set the IR filter ahead of time for determining
what key should be used to wake-up the device when in standby.  In doing so,
there is no longer any need for specifically enabling and disabling the IR
data filter when entering and exiting standby respectively.  As a consequence,
all the code in libnexusir dealing with placing the IR in to standby and
exiting standby has now been removed.

Please note that the refsw should have the changes from SW7445-2314 in order
for this to work properly (our Android refsw branch does).

refs #SWANDROID-1645

Change-Id: Ic88faf7ca652fb681b3dd62a9eeadf9b9afa1f03