nexusclient: get object refactoring.
slight refactor of the call, the passing of a null
object to attempt querying the number of objects to
return is an invalid refsw call and may only work out
of luck.  it certainly isn't working properly in
later refsw releases, so we remove it here all together
to unblock those.

refs #SWANDROID-1694

Change-Id: I09792d1489f70f60cc4fda62b7da46987c734dc6
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>