Generate stats on video decoder
In order to enable this feature, rebuild the
library libstagefrighthw.so with the flag
VDEC_STATS turned on. The decoder will print
to logcat some basic statistics about the playback
session such as bitrate, fps, number of dropped
frames, etc. It will also generate a text file
under /video/tombstones which will contain the
system time and presentation timestamp of all the
input/output/ frames. This could be used to track
decoder performance issues, a/v sync issues and
perhaps other problems.

refs #SWANDROID-1827

Change-Id: Idbc915907949d0dbd2d1995fae22f8c01a4da871
Signed-off-by: Luis Navarro <navarro@broadcom.com>
Signed-off-by: Edwin Chan <echan@broadcom.com>