SWANDROID-1702: [L] Allow custom override of power management and HDMI/CEC properties
Part 1 of 2 changes:
===================

This is the first set of changes to allow the power management and HDMI+CEC Android
properties from being customised per product.  The default property values are now
specified in the code and the "pre_settings.mk" file can override these values
using "LOCAL_PRODUCT_PROPERTY_OVERRIDES".  The second set of changes will include
those in the "device/broadcom" repo.

I've also removed the ANDROID_SUPPORTS_PM option, as we now always default to
having ANDROID PM enabled and it doesn't make sense otherwise to be able to
disable it.

Finally I've renamed the LOG TAG for HDMI to allow filtering of it to work
with logcat.

refs #SWANDROID-1702

Change-Id: I7891a696a58559a0a74a3cc715ff87b5d2607d36