Converting libb_os to use Android.mk directly
This converts libb_os to use an Android.mk file instead of building
externally and getting pulled in via brcm_nexus/Android.mk.  For now
since libstagefrighthw is the only user it's compiling the files
into that library but it's easily separated if we need to down the
road.  This seems to resolve issues with pthread_key_delete randomly
hanging during B_Os_Uninit.

refs #SWANDROID-1527

Change-Id: I1195fd104d9d1060d4c0bb94c2a27c122a745971