SWANDROID-1729: [L] SAGE crash when entering and exiting S5 quickly
We no longer wait 5s for the interactive timeout to expire before
beginning the shutdown process.  Instead, we immediately begin
the shutdown process and no longer call Nexus to place it in to
standby prior to shutting down.  When the platform powers off,
the Nexus driver's reboot_notifier will automatically be called
to place Nexus in to deep sleep (S5) and the platform will be in
a state ready to wakeup on the sources previously setup (e.g.
IR, CEC, BLE).

In making this change, there is no longer an opportunity for the
user to enter and exit S5 quickly via the POWER button on the
remote control.  Hence, no SAGE crash is ever observed.

The other advantage of this change, is that the system enters
deep sleep (S5) as quickly as it would when the user does a long
press of the POWER button and confirms that the STB should
shutdown.

refs #SWANDROID-1729

Change-Id: If5e21a9c9a188e2eedf7fce37c66c0882670bbc9