SWANDROID-1348: [L] Feature: STB should wake up on receipt of CEC commands.
These changes work in conjunction with SWANDROID-1597 refsw changes and
allow the <Set Stream Path> and
<User Control Pressed>["Power", "Power On", "Power Toggle"] CEC messages
to wake up the STB in any standby mode.  If the STB is in S1
standby, then an additional check is made to ensure that the CEC
message received is a valid wake-up message before forwarding it to
Android's HDMI Control Service.

refs #SWANDROID-1348

Change-Id: I385862c82344dd32965012883da62da366a7bf3a